\name{shinyPredictor}
\alias{shinyPredictor}
    shinyPredictor
\title{
    Predicting Diamond Prices
}
\description{
    This shiny app uses two predictors to estimate the price of a diamond
    based on the carat value and the cut type. More specifically, it had
    two linear regression model which use different variables to predict
    the price of the diamonds. You can choose a carat value and the cut type
    and check an estimate price from each predictor model. This estimate could
    be improved with more advanced non-linear predictors.
}
\usage{
    Use the sliders to select values for carat and cut type and click submit
    to get price predictions from the two linear models.
}
\author{
    Panayiotis L.
}
\note{
    This is a very simple shiny app with limited prediction capabilities.
}
\keyword{ ~shinyapp }
\keyword{ ~diamondsdata }
